implmentando funcao de map e renderizando as categorias
